+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.1
 
 - `[Tokens]` Added export in DTCG format with descriptions. Later this will be the primary export. For now its for testing and getting the values out of JSON. Note that it includes descriptions. ([4461](https://inforwiki.atlassian.net/browse/IDS-4461))
+- `[Tokens]` Changed `ids-dataviz-color` tokens to `ids-color-dataviz`. ([4575](https://inforwiki.atlassian.net/browse/IDS-4575))
 
 ## v1.1.0